Process scheduling showing waiting queue and cpu status as per gantt chart
For a given set of processes, their placing in the ready queue , waiting queue will be shown along with status of CPU. This will be shown as per the CPU gantt chart. For ready queue, multilevel feedback queue will be implemented (sjf in first queue, fcfs in second queue and longest job first in third queue). For each process, a unique ID will be given along with total I/O and computational runtime. Status of each process in both the queues are shown.